check file exist before copy

+++ res.sh
@@ -2,9 +2,27 @@
 echo $1 $2
 TARGET_PRODUCT=$1
 PRODUCT_OUT=$2
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/8188eu.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/8188eu.ko $PRODUCT_OUT/recovery/root/res/
+fi
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/8192cu.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/8192cu.ko $PRODUCT_OUT/recovery/root/res/
+fi
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/rda5890.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/rda5890.ko $PRODUCT_OUT/recovery/root/res/
+fi
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/rt5370sta.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/rt5370sta.ko $PRODUCT_OUT/recovery/root/res/
+fi
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/rkwifi.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/rkwifi.ko $PRODUCT_OUT/recovery/root/res/
+fi
+
+if [ -e "device/rockchip/$TARGET_PRODUCT/proprietary/libipp/rk29-ipp.ko" ] ; then
 cp device/rockchip/$TARGET_PRODUCT/proprietary/libipp/rk29-ipp.ko $PRODUCT_OUT/recovery/root/
+fi
